backend/requirements.txt
# ── Web framework ────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
asgi-correlation-id>=4.3.0
python-multipart>=0.0.9

# ── Database ──────────────────────────────────────────────────
sqlalchemy>=2.0.30
alembic>=1.13.1
asyncpg>=0.29.0

# ── Validation / config ───────────────────────────────────────
pydantic>=2.7.4
pydantic-settings>=2.3.4

# ── Auth / security ───────────────────────────────────────────
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.5
itsdangerous>=2.2.0
email-validator>=2.1.1

# ── AI / vector ───────────────────────────────────────────────
openai>=1.30.0
chromadb>=0.5.0
tiktoken>=0.7.0

# ── HTTP clients / integrations ───────────────────────────────
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.0

# ── Observability ─────────────────────────────────────────────
structlog>=24.2.0

# ── Dev / test ────────────────────────────────────────────────
black>=24.4.2
pytest>=8.2.2
pytest-asyncio>=0.23.7
pytest-mock>=3.14.0
